Share legacy-scripting helpers between get and render auth template
